pyproject.toml
[tool.poetry]
name = "bind-rest-api"
version = "0.1.0"
description = "Rest API for Bind9 built with FastAPI"
authors = ["Jay Tuckey <jay.tuckey@cdu.edu.au>"]
license = "GPL3"

[tool.poetry.dependencies]
python = "^3.8"
dnspython = "^2.0.0"
fastapi = "^0.61.1"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
